refactor flags
