+++ src/assembly/cpu/flags.ts
@@ -1,0 +1,10 @@
+export enum Flags {
+  N = 1 << 7,
+  V = 1 << 6,
+  U = 1 << 5,
+  B = 1 << 4,
+  D = 1 << 3,
+  I = 1 << 2,
+  Z = 1 << 1,
+  C = 1
+}
